Implemented population evolution for robustness.

+++ src/CGP.jl
@@ -9,7 +9,7 @@ perfectly suitable for other applications as well.
 
 using Combinatorics
 module CGP
-const MyInt = UInt32     # Type of bit string integers used in bit functions
+const MyInt = UInt16     # Type of bit string integers used in bit functions
 const MyFunc = UInt128  # Type of concatenated output representation of functions
 include("aliases.jl")
 include("Contexts.jl")
